# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Workforce OS v2 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: Paginated GraphRun history
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated graph runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRuns"
...
  /runs/{id}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Get full evidence timeline for one run
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Run with evidence timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphRunDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    GraphRun:
      type: object
      properties:
        id: { type: string }
        status:
          type: string
          enum: [RUNNING, AWAITING_APPROVAL, COMPLETED, FAILED]
        agentsInvolved:
          type: array
          items: { type: string }
        leadsSourced: { type: integer }
        artifactsGenerated: { type: integer }
        durationMs: { type: integer }
        costUsd: { type: number }
        triggeredBy: { type: string }
        startedAt: { type: string }
        completedAt: { type: ["string", "null"] }
      required: [id, status, agentsInvolved, leadsSourced, artifactsGenerated, durationMs, costUsd, triggeredBy, startedAt]
    PaginatedRuns:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GraphRun"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [items, total, page, limit]
    GraphRunDetail:
      type: object
      properties:
        run:
          $ref: "#/components/schemas/GraphRun"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineNode"
      required: [run, timeline]
...
    TimelineNode:
      type: object
      properties:
        id: { type: string }
        nodeType:
          type: string
          enum: [agent_run, llm_call, evaluator, tool_call, human_action]
        label: { type: string }
        summary: { type: string }
        reasoning: { type: ["string", "null"] }
        tokensUsed: { type: ["integer", "null"] }
        durationMs: { type: ["integer", "null"] }
        cost: { type: ["number", "null"] }
        score: { type: ["number", "null"] }
        timestamp: { type: string }
        children:
          type: array
          items:
            $ref: "#/components/schemas/TimelineNode"
      required: [id, nodeType, label, summary, timestamp, children]