# lib/api-zod/src/generated/types/graphRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphRunStatus } from './graphRunStatus';

export interface GraphRun {
  id: string;
  status: GraphRunStatus;
  agentsInvolved: string[];
  leadsSourced: number;
  artifactsGenerated: number;
  durationMs: number;
  costUsd: number;
  triggeredBy: string;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}