# lib/api-zod/src/generated/types/graphRunDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphRun } from './graphRun';
import type { TimelineNode } from './timelineNode';

export interface GraphRunDetail {
  run: GraphRun;
  timeline: TimelineNode[];
}